+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Added ability to change what EPP is linked with each throttle profile
+- Don't change EPP or thermal profile if the battery/ac state hasn't actually changed on resume
+- Re-implement the `asusctl -s` command (not fully)
+- Add more docs to some parts of code, and dbus interfaces
 
 ## [v5.0.7]
 ### Changed